Minor fixes across whole GEC module
